feat: add startChecking and stopChecking in lieu of check with intervals
