+++ lib/check.js
@@ -6,9 +6,7 @@ var internals = module.exports.internals = {}
 internals.nets = require('nets')
 internals.cache = require('./utils/cache')
 
-function check (state, options) {
-  options = parse(options)
-
+function check (state) {
   return new Promise(function (resolve, reject) {
     internals.nets({
       method: state.method,
@@ -20,7 +18,7 @@ function check (state, options) {
       if (error) {
         error.name = error.code === 'ETIMEDOUT' ? 'TimeoutError' : 'ConnectionError'
         error.code = undefined
-        handleError(state, error, options)
+        handleError(state, error)
         return reject(error)
       }
 
@@ -28,7 +26,7 @@ function check (state, options) {
         error = new Error('Server error')
         error.name = 'ServerError'
         error.code = response.statusCode
-        handleError(state, error, options)
+        handleError(state, error)
         return reject(error)
       }
 
@@ -38,41 +36,16 @@ function check (state, options) {
       }
 
       internals.cache.set(state)
-      handleInterval(state, options)
 
       resolve()
     })
   })
 }
 
-function parse (options) {
-  if (!options) {
-    return {}
-  }
-
-  return options
-}
-
-function handleError (state, error, options) {
+function handleError (state, error) {
   if (!state.error) {
     state.emitter.emit('disconnected')
   }
   state.error = error
   internals.cache.set(state, error)
-  handleInterval(state, options)
-}
-
-function handleInterval (state, options) {
-  if (options.interval) {
-    // we use setTimeout on purpose, we don't want to send requests each
-    // x seconds, but rather set a timeout for x seconds after each response
-    // but we use `interval` as variable as the effect is the same
-    state.interval = setTimeout(check, options.interval, state, options)
-    return
-  }
-
-  if (state.interval) {
-    clearTimeout(state.interval)
-    delete state.interval
-  }
 }
