+++ lib/stopChecking.js
@@ -1,0 +1,6 @@
+module.exports = stopChecking
+
+function stopChecking (state) {
+  clearTimeout(state.interval)
+  delete state.interval
+}
