+++ tests/unit/check.js
@@ -207,116 +207,3 @@ test('check() with 200 response when state.error is set', function (t) {
       simple.restore()
     })
 })
-
-test('check() with interval & 200 response', function (t) {
-  t.plan(2)
-
-  var clock = lolex.install(0)
-  var emitter = {emit: function () {}}
-
-  var state = {
-    method: 'HEAD',
-    url: 'https://example.com/ping',
-    emitter: emitter
-  }
-
-  simple.mock(check.internals.cache, 'set').callFn(function () {})
-  simple.mock(check.internals, 'nets').callbackWith(null, {
-    statusCode: 200
-  })
-
-  check(state, {
-    interval: 1000
-  })
-
-    .then(function () {
-      t.ok(state.interval, 'state.interval is set')
-      clock.tick(1000)
-      t.is(check.internals.nets.callCount, 2, '2 requests sent')
-    })
-
-    .catch(t.fail)
-
-    // cleanup
-    .then(function () {
-      clearTimeout(state.interval)
-      simple.restore()
-      clock.uninstall()
-    })
-})
-
-test('check() with interval & 500 error response', function (t) {
-  t.plan(2)
-
-  var clock = lolex.install(0)
-  var emitter = {emit: function () {}}
-
-  var state = {
-    method: 'HEAD',
-    url: 'https://example.com/ping',
-    emitter: emitter
-  }
-
-  simple.mock(check.internals.cache, 'set').callFn(function () {})
-  simple.mock(check.internals, 'nets').callbackWith(null, {
-    statusCode: 500
-  })
-
-  check(state, {
-    interval: 1000
-  })
-
-    .catch(function () {
-      t.ok(state.interval, 'state.interval is set')
-      clock.tick(1000)
-      t.is(check.internals.nets.callCount, 2, '2 requests sent')
-    })
-
-    // cleanup
-    .then(function () {
-      clearTimeout(state.interval)
-      simple.restore()
-      clock.uninstall()
-    })
-})
-
-test('check() with existing interval', function (t) {
-  t.plan(2)
-
-  var clock = lolex.install(0)
-  var emitter = {emit: function () {}}
-
-  var state = {
-    method: 'HEAD',
-    url: 'https://example.com/ping',
-    emitter: emitter,
-    interval: setTimeout(function () {
-      t.fail('timeout should not be called')
-    }, 1000)
-  }
-
-  setTimeout(function () {
-    t.pass('sanity check: 1000ms passed')
-  }, 1000)
-
-  simple.mock(check.internals.cache, 'set').callFn(function () {})
-  simple.mock(check.internals, 'nets').callbackWith(null, {
-    statusCode: 200
-  })
-
-  check(state)
-
-    .then(function () {
-      t.is(state.interval, undefined, 'state.interval has been removed')
-      clock.tick(1000)
-    })
-
-    .catch(t.fail)
-
-    // cleanup
-    .then(function () {
-      clearTimeout(state.interval)
-      simple.restore()
-      clock.uninstall()
-    })
-})
