+++ tests/unit/index.js
@@ -2,3 +2,5 @@ require('./check')
 require('./get-ok')
 require('./utils-cache')
 require('./utils-parse-options')
+require('./startChecking')
+require('./stopChecking')
